Dockerizado. La modificiación de ciertas líneas de código arremetió con el uso de venv, pero se confirma que desde el contenedor se ejecuta sin mayor novedad

+++ actividad4/src/models/entities/usuario.py
@@ -1,4 +1,4 @@
-from sqlalchemy import BigInteger, String
+from sqlalchemy import Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 from src import db
 
@@ -6,7 +6,7 @@ from src import db
 class Usuario(db.Model):
     __tablename__ = "usuarios"
     id: Mapped[int] = mapped_column(
-        BigInteger, primary_key=True, autoincrement=True
+        Integer, primary_key=True, autoincrement=True
     )
     username: Mapped[str] = mapped_column(
         String(80), unique=True, nullable=False
